Adding 3 new 'form' pattern pages (they haven't previously been on confluence)
The drafts from GOV.UK showed markup rather than examples. I'm in two minds whether to use the codesnippet boxes or example boxes for the two pages with examples. For lots of small markup examples with no styling the example boxes make the pages very long and add an extra click for users. Also some of the examples are more suited to being shown are markup. But they are a clearer way to show the text.