Update package-lock.json
Previous package-lock.json was probably created with a version of node
and npm than the node version used by jenkins, which meant that when
jenkins ran an npm install as part of the build it caused the
package-lock.json to change and that caused the build to fail. Fixed by
running `nvm install && npm install` which installed the version of node
from the .nvmrc file listed in the project then ran npm install again to
regenerate the package-lock.json file.